refactor: split cosmic impl
