+++ src/app/nav.rs
@@ -1,0 +1,26 @@
+use cosmic::{
+    app::{self},
+    widget::{self},
+    ApplicationExt, Task,
+};
+
+use crate::app::message::Message;
+use crate::app::page::Page;
+use crate::app::App;
+use crate::fl;
+
+use super::Cosmic;
+
+impl Cosmic {
+    pub fn on_nav_select(app: &mut App, id: widget::nav_bar::Id) -> app::Task<Message> {
+        app.cosmic.nav_model.activate(id);
+
+        let title = if let Some(page) = app.cosmic.nav_model.data::<Page>(id) {
+            format!("{} - {}", page.title(), fl!("app-title"))
+        } else {
+            fl!("app-title")
+        };
+
+        Task::batch(vec![app.set_window_title(title)])
+    }
+}
