+++ src/app/subscription.rs
@@ -1,0 +1,66 @@
+use std::any::TypeId;
+
+use cosmic::{
+    cosmic_config::{self, Update},
+    cosmic_theme::{self, ThemeMode},
+    iced::{event, keyboard::Event as KeyEvent, Event, Subscription},
+    Application,
+};
+
+use crate::app::message::Message;
+use crate::app::App;
+
+use crate::core::config::CONFIG_VERSION;
+
+use super::Cosmic;
+
+impl Cosmic {
+    pub fn subscription() -> cosmic::iced::Subscription<Message> {
+        struct ConfigSubscription;
+        struct ThemeSubscription;
+
+        let subscriptions = vec![
+            event::listen_with(|event, _status, _window_id| match event {
+                Event::Keyboard(KeyEvent::KeyPressed { key, modifiers, .. }) => {
+                    Some(Message::Key(modifiers, key))
+                }
+                Event::Keyboard(KeyEvent::ModifiersChanged(modifiers)) => {
+                    Some(Message::Modifiers(modifiers))
+                }
+                _ => None,
+            }),
+            cosmic_config::config_subscription(
+                TypeId::of::<ConfigSubscription>(),
+                App::APP_ID.into(),
+                CONFIG_VERSION,
+            )
+            .map(|update: Update<ThemeMode>| {
+                if !update.errors.is_empty() {
+                    log::info!(
+                        "errors loading config {:?}: {:?}",
+                        update.keys,
+                        update.errors
+                    );
+                }
+                Message::SystemThemeModeChange
+            }),
+            cosmic_config::config_subscription::<_, cosmic_theme::ThemeMode>(
+                TypeId::of::<ThemeSubscription>(),
+                cosmic_theme::THEME_MODE_ID.into(),
+                cosmic_theme::ThemeMode::version(),
+            )
+            .map(|update: Update<ThemeMode>| {
+                if !update.errors.is_empty() {
+                    log::info!(
+                        "errors loading theme mode {:?}: {:?}",
+                        update.keys,
+                        update.errors
+                    );
+                }
+                Message::SystemThemeModeChange
+            }),
+        ];
+
+        Subscription::batch(subscriptions)
+    }
+}
